Added SSH tunnel config
