+++ keychain.js
@@ -1,0 +1,16 @@
+const oauth = require("./client/src/config/oauth");
+
+var keychain = {
+    "username": "admin",
+    "keyPath": "/home/classinsight/.ssh/id_rsa",
+    "password": "youdownwithrpp",
+    "dbHost": "edusense-dev-1.andrew.cmu.edu",
+    "sso": {
+        "google": {
+            "clientId": oauth.clientId,
+            "clientSecret": oauth.clientSecret
+        }
+    }
+}
+
+module.exports = keychain;
